Moved DafnyDocument and DiagnosticPublisher to the Workspace namespace
