+++ Source/DafnyLanguageServer/DafnyRequestNames.cs
@@ -2,5 +2,6 @@
   public static class DafnyRequestNames {
     public const string CounterExample = "dafny/counterExample";
     public const string CompilationStatus = "dafny/compilation/status";
+    public const string GhostDiagnostics = "dafny/ghost/diagnostics";
   }
 }
