+++ Source/DafnyLanguageServer/Util/PathExtensions.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.Dafny.LanguageServer.Language;
+﻿using Microsoft.Dafny.LanguageServer.Workspace;
 using OmniSharp.Extensions.LanguageServer.Protocol;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using System.IO;
