+++ Source/DafnyLanguageServer/Workspace/DocumentDatabase.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.Boogie;
-using Microsoft.Dafny.LanguageServer.Language;
 using Microsoft.Dafny.LanguageServer.Workspace.ChangeProcessors;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
