+++ Source/DafnyLanguageServer/Workspace/IDocumentDatabase.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.Dafny.LanguageServer.Language;
-using OmniSharp.Extensions.LanguageServer.Protocol.Models;
+﻿using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using System;
 using System.Threading.Tasks;
