+++ Source/DafnyLanguageServer/Workspace/LanguageServerExtensions.cs
@@ -32,7 +32,8 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
         .AddSingleton<ITextChangeProcessor, TextChangeProcessor>()
         .AddSingleton<ISymbolTableRelocator, SymbolTableRelocator>()
         .AddSingleton<ISymbolGuesser, SymbolGuesser>()
-        .AddSingleton<ICompilationStatusNotificationPublisher, CompilationStatusNotificationPublisher>();
+        .AddSingleton<ICompilationStatusNotificationPublisher, CompilationStatusNotificationPublisher>()
+        .AddSingleton<IDiagnosticPublisher, DiagnosticPublisher>();
     }
 
     private static TextDocumentLoader CreateTextDocumentLoader(IServiceProvider services) {
